Fix some bugs.
Need a lock to protect refcnts on the global config structures.

evl_peek would return the wrong value on an empty ring.

Change thread semantics so the destroy function doesn't get the evpl agent reference, as cleanup is simpler if we assume its already been destroyed by then.

Give the expected full length of inner RPC2 messages to the generated unmarshall code so it can check if they match.